ShopperUser Trait WIP
